Add support for building messages with LibXML
Created RSolr::Message::Builders namespace, which contains a Builder
class and a Libxml class. These classes implement the same interface for
building XML messages, using the respective XML libraries for which they
are named. The RSolr::Message singleton contains the Builder
implementation by default, but has a cattr_accessor to change it to the
Libxml implementation.