Add basic JSON building capabilities.
The main piece of functionality lacking at this point is multi-value
fields. Currently the XML duplicates them. Perhaps it's worth
subclassing document for XML and JSON so that we get the main document
logic exposed as well as proper management of these fields.

Aside from that, the only thing remaining is to have a way of choosing
to add using json rather than XML.