Pass `timeout` arg directly to faraday, deprecate `read_timeout`
Since RSolr 2.0, the `read_timeout` option given to Rsolr.connect is translated to a `timeout` option passed to the underlying Faraday object.

This makes a confusing situation even more confusing. Why not just take and pass a `timeout` option as `timeout` to Faraday? This PR does that. For backwards compatibilty, if you pass read_timeout, it will still also be passed as Faraday `timeout` (same as it was before this PR), but a deprecation message will be printed out. (Using `warn`, which the codebase already has one instance of using for a deprecation warning, so I just followed suit).

Really for at least some adapters, Faraday supports all of: `timeout`, `read_timeout`, `write_timeout`, and `open_timeout`. Ideally we might allow and pass ALL of these on to Faraday... but there's no great way to do that backwards compatibly when previous RSolr took `read_timeout` and passed it as `timeout`! So one step at a time, don't change what functionality is supported, just make the name direct, as a stepping stone to supporting more Faraday timeout options.

There were no tests for this behavior previously. I have added no tests. I could not figure out any reasonable way to write a test here; if anyone has one, suggestions welcome. There is no drop in test coverage, but this is not ideal.